// src/app/core/modele/widok-mapy-mapa-szczegoly-dto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WidokMapyPunktDto } from './widok-mapy-punkt-dto';
import { WidokMapyJezykDto } from './widok-mapy-jezyk-dto';
import { WidokMapyDomyslneWarstwyPodkladowe } from './widok-mapy-domyslne-warstwy-podkladowe';
import { WidokMapyWarstwaDto } from './widok-mapy-warstwa-dto';


export interface WidokMapyMapaSzczegolyDto {
    uuid?: string;
    nazwa?: WidokMapyJezykDto;
    opisWPortaluMapowym?: WidokMapyJezykDto;
    domyslnaSkala?: number;
    srodekMapyX?: number;
    srodekMapyY?: number;
    domyslnaSkalaWyszukiwania?: number;
    domyslneWarstwyPodkladowe?: WidokMapyDomyslneWarstwyPodkladowe;
    warstwy: Array<WidokMapyWarstwaDto>;
    limitWarstw?: number;
    udostepnionaLokalizacja?: WidokMapyPunktDto;
    usunieteWarstwy?: Array<WidokMapyJezykDto>;
}